add policy for task and commonLife page

+++ app/Http/Controllers/CommonLifeController.php
@@ -4,14 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\task;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class CommonLifeController extends Controller
 {
+    use AuthorizesRequests;
     public function index() {
         $task = Task::all();
         return view('pages.commonLife.index', compact('task'));
     }
     public function store(Request $request) {
+        $this->authorize('create', Task::class);
         $task=Task::Create([
             'title' => $request->input('title'),
             'description' => $request->input('description'),
@@ -21,18 +24,20 @@ class CommonLifeController extends Controller
 
     public function delete($id) {
         $task=Task::findOrFail($id);
+        $this->authorize('delete', $task);
         $task->delete();
         return redirect()->back();
     }
 
     public function update(Request $request, $id)
     {
+        $task = Task::findOrFail($id);
+        $this->authorize('update', $task);
         $request->validate([
             'titleEdit' => 'required|string|max:255',
             'descriptionEdit' => 'required|string',
         ]);
 
-        $task = Task::findOrFail($id);
         $task->title = $request->input('titleEdit');
         $task->description = $request->input('descriptionEdit');
         $task->save();
